feat: separated types to different pkg
