+++ packages/constants/src/index.ts
@@ -6,18 +6,13 @@ export * from './lib/constants/constants';
 export * from './lib/constants/defaultLitNodeClientConfig';
 
 // ----------- Interfaces -----------
-export * from './lib/interfaces/interfaces';
 export * from './lib/interfaces/i-errors';
-export * from './lib/interfaces/ILitNodeClient';
 
 // ----------- ENUMS -----------
 export * from './lib/enums';
 
 // ----------- Errors -----------
 export * from './lib/errors';
-
-// ----------- Types -----------
-export * from './lib/types';
 
 // ----------- Utils -----------
 export * from './lib/utils/utils';
