+++ packages/encryption/src/lib/params-validators.ts
@@ -4,6 +4,10 @@
  */
 
 import {
+  LIT_ERROR,
+} from '@lit-protocol/constants';
+
+import {
   DecryptFileProps,
   JsonEncryptionRetrieveRequest,
   JsonSaveEncryptionKeyRequest,
@@ -11,8 +15,8 @@ import {
   EncryptFileAndZipWithMetadataProps,
   KV,
   ExecuteJsProps,
-  LIT_ERROR,
-} from '@lit-protocol/constants';
+} from '@lit-protocol/types';
+
 import {
   checkIfAuthSigRequiresChainParam,
   checkType,
