+++ packages/types/README.md
@@ -1,0 +1,19 @@
+# types
+
+This library was generated with [Nx](https://nx.dev).
+
+
+
+## Building
+
+Run `nx build types` to build the library.
+
+
+
+
+
+## Running unit tests
+
+Run `nx test types` to execute the unit tests via [Jest](https://jestjs.io).
+
+
