+++ packages/types/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@lit-protocol/types",
+  "version": "2.1.55",
+  "type": "commonjs",
+  "license": "MIT",
+  "homepage": "https://github.com/Lit-Protocol/js-sdk",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/LIT-Protocol/js-sdk"
+  },
+  "keywords": [
+    "library"
+  ],
+  "bugs": {
+    "url": "https://github.com/LIT-Protocol/js-sdk/issues"
+  },
+  "publishConfig": {
+    "access": "public",
+    "directory": "../../dist/packages/types"
+  },
+  "tags": [
+    "universal"
+  ],
+  "main": "./dist/src/index.js",
+  "typings": "./dist/src/index.d.ts"
+}
