+++ workspace.json
@@ -22,6 +22,7 @@
     "nodejs": "apps/nodejs",
     "pkp-ethers": "packages/pkp-ethers",
     "react": "apps/react",
+    "types": "packages/types",
     "uint8arrays": "packages/uint8arrays"
   }
 }
